requirements.txt
annotated-types==0.6.0
anyio==4.6.2
appnope==0.1.4
asttokens==3.0.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
comm==0.2.2
cryptography==44.0.2
debugpy==1.8.11
decorator==5.2.1
# Removed deprecated 'dotenv==0.9.9'; using python-dotenv instead.
Events==0.5
exceptiongroup==1.2.0
executing==2.1.0
fastapi==0.112.2
greenlet==3.1.1
idna==3.7
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.34.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest_asyncio==1.6.0
opensearch==0.9.2
opensearch-py==2.8.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.2.0
requests==2.32.3
setuptools==75.8.0
six==1.17.0
sniffio==1.3.0
SQLAlchemy==2.0.37
stack_data==0.6.3
starlette==0.38.2
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
wheel==0.45.1
zipp==3.21.0
uvicorn==0.32.1
email_validator==2.2.0

